pran.ai/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
fastapi
uvicorn
pydantic
python-dotenv
requests
openai